customer_support_env/requirements.txt
openenv-core>=0.2.3
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
gradio>=4.20.0
openai>=1.0.0
numpy>=1.26.0
pytest>=8.0.0
